Implement UDP as a service
UDP Services are notoriously hard to keep track of, as UDP is a
stateless protocol. The servers do *not* have to respond to requests,
unlike TCP with its three-way handshake.

However, we can use ICMP protocol (as do netcat and nmap) to identify
running services. If the service is offline, we get 'ICMP port
unreachable' instead of no response.

Using ICMP is *not* an exact guarantee, as other factors (like firewall)
can also lead to the same result. Still, it's better than nothing.

We also have to run Puptime as root, since creating ICMP packets
requires access to "raw" sockets.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>