Add log file, pid file as config variables
As we are using ICMP to keep track of UDP services, we have to run the
Puptime in root. As such, expecting log and pid files to be placed in
'~/.puptime/' is imposing. Let's allow log file and pid file to be
defined in config file and use '~/.puptime/server.log' and
'~/.puptime/pid.log' as defaults.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>