Implement teams notification
We are going to use
[`msteams-ruby-client`](https://github.com/oooooooo/msteams-ruby-client)
to help get the project working as soon as possible. Unlike other
options, the wrapper library uses channel-specific webhook end points to
communicate.

These endpoints can be enabled as described here: [Add an incoming
webhook to a teams
channel](https://docs.microsoft.com/en-us/microsoftteams/platform/webhooks-and-connectors/how-to/add-incoming-webhook#add-an-incoming-webhook-to-a-teams-channel).
Once enabled, these webhook urls must be kept private.

The communication between Puptime and the webhook is rather simple: We
create a message with title "Puptime Notification" and content as
`@message`.

Signed-off-by: Abhishek Kumar <abhishekkumar8222@gmail.com>